+++ crates/wysiwyg/src/tests/test_links.rs
@@ -15,7 +15,7 @@
 use crate::tests::testutils_composer_model::{cm, tx};
 use crate::tests::testutils_conversion::utf16;
 
-use crate::{LinkAction, TextUpdate};
+use crate::TextUpdate;
 
 #[test]
 fn cant_set_link_to_empty_selection() {
@@ -322,25 +322,4 @@ fn replace_text_in_a_link_inside_a_list_partially_selected_starting_inside_endin
     let mut model = cm("<ul><li>list_element</li><li><a href=\"https://element.io\">link{_in_}|list</a></li></ul>");
     model.replace_text(utf16("added_text"));
     assert_eq!(tx(&model), "<ul><li>list_element</li><li><a href=\"https://element.io\">linkadded_text|list</a></li></ul>");
-}
-
-#[test]
-fn link_action_from_non_highlighted_normal_text() {
-    let model = cm("test|");
-    assert_eq!(model.get_link_action(), LinkAction::CreateWithText)
-}
-
-#[test]
-fn link_action_from_highlighted_normal_text() {
-    let model = cm("{test}|");
-    assert_eq!(model.get_link_action(), LinkAction::Create)
-}
-
-#[test]
-fn link_action_from_highlighted_link() {
-    let model = cm("{<a href=\"https://element.io\">test</a>}|");
-    assert_eq!(
-        model.get_link_action(),
-        LinkAction::Edit(utf16("https://element.io"))
-    )
 }
